@@ -1,22 +1,38 @@
 {
-  "name": "uqa-frontend",
-  "version": "0.0.1",
-  "description": "",
+  "name": "uQA-frontend",
+  "version": "1.0.0",
+  "description": "Web frontend for uQA",
   "main": "js/Main.js",
-  "scripts": {
-    "dev": "node dist/app/index.js",
-    "test": "mocha -r should -R spec"
+  "dependencies": {
+    "vue": "^2.1.4",
+    "vue-resource": "^1.0.3",
+    "vue-router": "^2.1.1"
   },
-  "author": "Jared Loomis",
-  "license": "ISC",
-  "dependencies": {},
   "devDependencies": {
+    "babel-core": "^6.18.2",
+    "babel-loader": "^6.2.9",
+    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-latest": "^6.16.0",
     "css-loader": "^0.26.1",
     "eslint-loader": "^1.6.1",
-    "file-loader": "^0.9.0",
+    "gulp": "^3.9.1",
+    "gulp-babel": "^6.1.2",
+    "gulp-util": "^3.0.7",
     "html-loader": "^0.4.4",
-    "webpack": "^2.1.0-beta.28",
+    "jshint": "^2.9.4",
+    "jshint-loader": "^0.8.3",
     "vue-loader": "^10.0.2",
-    "vue-template-compiler": "^2.1.6"
-  }
+    "vue-template-compiler": "^2.1.4",
+    "webpack": "^2.2.0",
+    "webpack-bundle-analyzer": "^2.2.1",
+    "webpack-dev-server": "^2.1.0-beta.0",
+    "webpack-stream": "^3.2.0"
+  },
+  "scripts": {
+    "dev": "webpack-dev-server --open --inline --hot",
+    "build": "webpack && npm run build-static",
+    "build-static": "cp view/Main.html dist/view/Main.html"
+  },
+  "author": "Jared Loomis",
+  "license": "All Rights Reserved"
 }
