@@ -1,9 +1,9 @@
-const passport = require("koa-passport")
+const passport      = require("koa-passport")
 const LocalStrategy = require("passport-local").Strategy
 
 const User = require("../model/User")
 
-module.exports = (app) => {
+module.exports = app => {
     passport.serializeUser((user, cb) => cb(null, user._id))
     passport.deserializeUser((id, cb) => {
         User.findById(id).exec().then((user) => cb(null, user))
@@ -11,14 +11,6 @@ module.exports = (app) => {
 
     const opts = {usernameField: "email"}
     passport.use(new LocalStrategy(opts, (email, password, cb) => {
-        User.find({email: email}).exec().then((user) => cb(null, user))
-        /*
-        User.findOne({email: email}).exec((err, user) => {
-            if(err) {
-                return cb(err, false)
-            }
-            cb(null, user)
-        })
-        */
+        User.find({email: email}).exec().then(user => cb(null, user))
     }))
 }
