@@ -1,9 +1,14 @@
-/* global describe, it */
 const jsc = require("jsverify")
+import test from "ava"
 
+test("Natural Numbers", t => {
+
+})
+
+/*
 describe("natural numbers", () => {
     it("are greater than or equal to 0", () => {
-        const property = jsc.forall("nat", (n) => {
+        const property = jsc.forall("nat", n => {
             return n >= 0
         })
 
@@ -11,10 +16,11 @@ describe("natural numbers", () => {
     })
 
     it("are less than 90", () => {
-        const property = jsc.forall("nat", (n) => {
+        const property = jsc.forall("nat", n => {
             return n < 90
         })
 
         jsc.assert(property, { size: 100 })
     })
 })
+*/
